+++ api/v1/cluster_types.go
@@ -24,7 +24,6 @@ import (
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/cloudnative-pg/cloudnative-pg/internal/configuration"
@@ -1128,19 +1127,6 @@ type StorageConfiguration struct {
 	// Template to be used to generate the Persistent Volume Claim
 	// +optional
 	PersistentVolumeClaimTemplate *corev1.PersistentVolumeClaimSpec `json:"pvcTemplate,omitempty"`
-}
-
-// TryParseQuantity returns the parsed quantity or nil
-func (storageConfiguration *StorageConfiguration) TryParseQuantity() *resource.Quantity {
-	if storageConfiguration == nil {
-		return nil
-	}
-
-	quantity, err := resource.ParseQuantity(storageConfiguration.Size)
-	if err != nil {
-		return nil
-	}
-	return &quantity
 }
 
 // SyncReplicaElectionConstraints contains the constraints for sync replicas election.
