+++ controllers/suite_test.go
@@ -45,6 +45,7 @@ import (
 	// +kubebuilder:scaffold:imports
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/certs"
+	pvcReconciler "github.com/cloudnative-pg/cloudnative-pg/pkg/reconciler/pvc"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 
@@ -337,7 +338,7 @@ func generateFakePVC(c client.Client, cluster *apiv1.Cluster) []corev1.Persisten
 	for idx < cluster.Spec.Instances {
 		idx++
 
-		pvc, err := specs.CreatePVC(cluster.Spec.StorageConfiguration, *cluster, idx, utils.PVCRolePgData)
+		pvc, err := pvcReconciler.Create(cluster.Spec.StorageConfiguration, *cluster, idx, utils.PVCRolePgData)
 		Expect(err).To(BeNil())
 		SetClusterOwnerAnnotationsAndLabels(&pvc.ObjectMeta, cluster)
 
@@ -345,7 +346,7 @@ func generateFakePVC(c client.Client, cluster *apiv1.Cluster) []corev1.Persisten
 		Expect(err).To(BeNil())
 		pvcs = append(pvcs, *pvc)
 		if cluster.ShouldCreateWalArchiveVolume() {
-			pvcWal, err := specs.CreatePVC(cluster.Spec.StorageConfiguration, *cluster, idx, utils.PVCRolePgWal)
+			pvcWal, err := pvcReconciler.Create(cluster.Spec.StorageConfiguration, *cluster, idx, utils.PVCRolePgWal)
 			Expect(err).To(BeNil())
 			SetClusterOwnerAnnotationsAndLabels(&pvcWal.ObjectMeta, cluster)
 			err = c.Create(context.Background(), pvcWal)
