+++ internal/cmd/plugin/destroy/destroy.go
@@ -29,7 +29,7 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/controllers"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/cmd/plugin"
 	"github.com/cloudnative-pg/cloudnative-pg/internal/plugin/resources"
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
+	pvcReconciler "github.com/cloudnative-pg/cloudnative-pg/pkg/reconciler/pvc"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 )
 
@@ -54,7 +54,7 @@ func Destroy(ctx context.Context, clusterName, instanceID string, keepPVC bool) 
 			}
 
 			pvcs[i].OwnerReferences = removeOwnerReference(pvcs[i].OwnerReferences, clusterName)
-			pvcs[i].Annotations[specs.PVCStatusAnnotationName] = specs.PVCStatusDetached
+			pvcs[i].Annotations[pvcReconciler.StatusAnnotationName] = pvcReconciler.StatusDetached
 			pvcs[i].Labels[utils.InstanceNameLabelName] = instanceName
 			err = plugin.Client.Update(ctx, &pvcs[i])
 			if err != nil {
@@ -75,7 +75,7 @@ func Destroy(ctx context.Context, clusterName, instanceID string, keepPVC bool) 
 		// but it does have the instance label and the detached annotation then we can still delete it
 		// We will only skip the iteration and not delete the pvc if it is not owned by the cluster, and it does not have
 		// the annotation or label
-		if isOwned || (pvcs[i].Annotations[specs.PVCStatusAnnotationName] == specs.PVCStatusDetached &&
+		if isOwned || (pvcs[i].Annotations[pvcReconciler.StatusAnnotationName] == pvcReconciler.StatusDetached &&
 			pvcs[i].Labels[utils.InstanceNameLabelName] == instanceName) {
 			if err = plugin.Client.Delete(ctx, &pvcs[i]); err != nil {
 				return fmt.Errorf("error deleting pvc %s: %v", pvcs[i].Name, err)
