+++ internal/plugin/resources/instance.go
@@ -35,6 +35,7 @@ import (
 	"github.com/cloudnative-pg/cloudnative-pg/internal/cmd/plugin"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/log"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/postgres"
+	pvcReconciler "github.com/cloudnative-pg/cloudnative-pg/pkg/reconciler/pvc"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 )
@@ -141,7 +142,7 @@ func GetInstancePVCs(
 
 	var pvcs []v1.PersistentVolumeClaim
 
-	pgDataName := specs.GetPVCName(cluster, instanceName, utils.PVCRolePgData)
+	pgDataName := pvcReconciler.GetPVCName(cluster, instanceName, utils.PVCRolePgData)
 	pgData, err := getPVC(ctx, pgDataName)
 	if err != nil {
 		return nil, err
@@ -150,7 +151,7 @@ func GetInstancePVCs(
 		pvcs = append(pvcs, *pgData)
 	}
 
-	pgWalName := specs.GetPVCName(cluster, instanceName, utils.PVCRolePgWal)
+	pgWalName := pvcReconciler.GetPVCName(cluster, instanceName, utils.PVCRolePgWal)
 	pgWal, err := getPVC(ctx, pgWalName)
 	if err != nil {
 		return nil, err
